fix(runtime): fix service runtime to stop abusing SD card write cycle
Previously the temp image from the camera was stored on the filesystem
which means it would get written to the SD card (assuming the OS
is running on the SD card). New version should mount a temp folder
into the virtual memory (ramfs) and store the temp image there. The ramfs
is used to prevent the storage to go into the swap which would result
in the disk abuse again.