fix(runtime): fix service runtime to stop abusing SD card write cycle (#2)
Previously the temp image from the camera was stored on the filesystem
which means it would get written to the SD card (assuming the OS is
running on the SD card).

The new version should mount a temp folder into the virtual memory
(`ramfs`) and store the temp image there.

The `ramfs` is used to prevent the storage from going into the swap
which would result in the permanent storage `I/O` abuse again.

While doing this I have also moved most of the `stdout` and `stderr` of
the individual scripts out from the service journal log again to reduce
the permanent storage `I/O` and moved the log files into the mounted
`ramfs` directory. Keeping only a single message that indicates the
script is failing and the user should consult logs. This was done again
in order to prevent excessive `I/O` into the `journalctl` log